requirements.txt
astroid==2.0.4
cycler==0.10.0
Django==2.1.3
django-filter==2.0.0
djangorestframework==3.9.0
isort==4.3.4
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Markdown==3.0.1
mccabe==0.6.1
pylint==2.1.1
pyparsing==2.3.0
pyserial==3.4
python-dateutil==2.7.5
pytz==2018.7
six==1.11.0
typed-ast==1.1.0
wrapt==1.10.11
